requirements.txt
numpy>=1.16.1
Keras>=3.7.2
gunicorn>=20.1.0
Flask>=2.2.3
tensorflow>=2.11.1
pillow>=9.4.0